feat(tsoa-setup): Renamed authentication middleware
